requirements.txt
playwright
langchain
langchain-anthropic
langgraph
anthropic
python-dotenv
markdownify
